+++ ChangeLog.rst
@@ -1,3 +1,9 @@
+HEAD
+====
+  - Bugfix: _struct_ref.pdbx_seq_one_letter_code is now treated as the subset
+    of the reference (e.g. UniProt) sequence that overlaps with our Entities,
+    not the entire sequence (#64).
+
 0.22 - 2021-10-22
 =================
   - The :class:`ihm.Software` class now allows a citation for the software